backend/requirements.txt
# Django 4+ & DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3.0
django-filter>=23.5

# MySQL
mysqlclient>=2.2.0

# JWT
djangorestframework-simplejwt>=5.3.0

# Redis & Celery
redis>=5.0.0
django-redis>=5.4.0
celery[redis]>=5.3.0
django-celery-beat>=2.5.0

# Security & validation
PyJWT>=2.8.0
phonenumbers>=8.13.0
python-dotenv>=1.0.0

# Utils
drf-spectacular>=0.27.0
